add Thread Connection functionnality

+++ src/main/java/com/excilys/om/Company.java
@@ -6,19 +6,24 @@ public class Company {
 	public String toString() {
 		return "Company [id=" + id + ", name=" + name + "]";
 	}
-	private long id;
+
+	private int id;
 	private String name;
-	public long getId() {
+
+	public int getId() {
 		return id;
 	}
-	public void setId(long id) {
+
+	public void setId(int id) {
 		this.id = id;
 	}
+
 	public String getName() {
 		return name;
 	}
+
 	public void setName(String name) {
 		this.name = name;
-	}	
-	
+	}
+
 }
