+++ src/main/java/com/excilys/om/Computer.java
@@ -4,48 +4,57 @@ import java.sql.Date;
 
 public class Computer {
 
-	private long id;
-	private String name;	
+	private int id;
+	private String name;
 	private Date introduced;
 	private Date discontinued;
 	private Company company;
-	
+
 	public String getName() {
 		return name;
 	}
+
 	public void setName(String name) {
 		this.name = name;
 	}
+
 	public Date getIntroduced() {
 		return introduced;
 	}
+
 	public void setIntroduced(Date introduced) {
 		this.introduced = introduced;
 	}
+
 	public Date getDiscontinued() {
 		return discontinued;
 	}
+
 	public void setDiscontinued(Date discontinued) {
 		this.discontinued = discontinued;
 	}
 
-	public long getId() {
+	public int getId() {
 		return id;
 	}
-	public void setId(long id) {
+
+	public void setId(int id) {
 		this.id = id;
 	}
+
 	public Company getCompany() {
 		return company;
 	}
+
 	public void setCompany(Company company) {
 		this.company = company;
 	}
+
 	@Override
 	public String toString() {
 		return "Computer [id=" + id + ", name=" + name + ", introduced="
 				+ introduced + ", discontinued=" + discontinued + ", company="
 				+ company + "]";
 	}
-	
+
 }
